Move File-based MultiPassStrategy from factory method into own class
